EH-132: Rename components
