+++ classification/unit_tests/CMakeLists.txt
@@ -1,5 +1,5 @@
-add_executable(read_classifier_test read_classifier_test.cc)
+add_executable(mapping_classifier_test mapping_classifier_test.cc)
 
-target_link_libraries(read_classifier_test classification gtest_main)
+target_link_libraries(mapping_classifier_test classification gtest_main)
 
-add_test(NAME read_classifier_test COMMAND read_classifier_test)
+add_test(NAME mapping_classifier_test COMMAND mapping_classifier_test)
