+++ graphs/unit_tests/graph_mapping_test.cc
@@ -184,12 +184,12 @@ TEST(GettingIndexesOfNode, NodeNotInMapping_EmptyListReturned) {
   EXPECT_EQ(empty_list, mapping.GetIndexesOfNode(4));
 }
 
-TEST(CheckingIfMappingSpansNode, TypicalMapping_ChecksPerformed) {
+TEST(CheckingIfMappingOverlapsNode, TypicalMapping_ChecksPerformed) {
   Graph graph = MakeStrGraph("AAAACC", "CCG", "ATTT");
   const string read = "ACCCCG";
   GraphMapping mapping = DecodeFromString(3, "0[3M]1[3M]", read, graph);
-  EXPECT_TRUE(mapping.SpansNode(0));
-  EXPECT_TRUE(mapping.SpansNode(1));
-  EXPECT_FALSE(mapping.SpansNode(2));
-  EXPECT_FALSE(mapping.SpansNode(3));
+  EXPECT_TRUE(mapping.OverlapsNode(0));
+  EXPECT_TRUE(mapping.OverlapsNode(1));
+  EXPECT_FALSE(mapping.OverlapsNode(2));
+  EXPECT_FALSE(mapping.OverlapsNode(3));
 }
